requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==4.2.11
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.1
psycopg2==2.9.9
PyJWT==2.8.0
redis==5.0.3
sqlparse==0.4.4
typing_extensions==4.11.0
